Update debian control file to depend on either grub-efi-i386 OR grub-efi-amd64
Since adca4024 the debian control file depends on either grub-pc or grub-efi, however currently in Ubuntu grub-efi *only* depends on grub-efi-amd64 but not grub-efi-i386, which could cause installation issues on i386-efi users.

This commits changes the dependency to depend on grub-pc OR grub-efi-amd64 OR grub-efi-i386, to fix this problem, unfortunately I don't have any 32-bit UEFI system to test this patch.

Signed-off-by: 林博仁 <Buo.Ren.Lin@gmail.com>