src/MainPanel.cpp - redirect error stream to /dev/null, fix possible off-by-N error, fixes #57
Currently the error string are inserted as an item of the target device list, which may causes winusbgui selecting the wrong target device to --format because of off-by-N error.

This commit fixes the problem by redirect error stream to /dev/null, refer popen(3) manpage.

This is considered a critical bug, however I'm lacking the knowledge to inspect what's going on in a C++ program memory using GDB so I'm also not sure if this patch is appropriate(or just making the problem even bigger), please review.

Signed-off-by: 林博仁 <Buo.Ren.Lin@gmail.com>