Debian packaging: Replace arch-dependent grub packages with grub-common
The previous problem of the dependency is that grub has ALOT of target architectures, and if running on an architecture that is not enumerated here APT will try to install grub-pc, which effectively removed the proper GRUB bootloader previously installed on system and interrupts the following bootup process.

This patch drops the dependency as it is not required anyway, we now depend on grub-common which shoud provide everything we need.

Signed-off-by: 林博仁 <Buo.Ren.Lin@gmail.com>