Fix missing file system detection in --partition preparation mode
Since commit 29d74ee(v3.1.5) a detection is implemented to prematurely
detect for files over 4GiB in the source image (the FAT32 single file
size limit), however it depends on the `target_filesystem_type` variable
to determine whether the check should be run, which isn't set according
to the existing target partition's file system in --partition
preparation mode, and thus the check will be run when the user's target
file system is NTFS and chooses to use the --partition creation mode
because the `target_filesystem_type` variable is set to FS_FAT by
default and isn't changed accordingly.

This patch implements the missing detection in
determine_target_parameters.

Fixes #226.

Signed-off-by: 林博仁(Buo-ren, Lin) <Buo.Ren.Lin@gmail.com>